+++ example/lint.sh
@@ -46,7 +46,7 @@ args+=(
 	# See https://github.com/aspect-build/rules_ts/pull/574#issuecomment-2073632879
 	"--norun_validations"
 	"--build_event_json_file=$buildevents"
-	"--output_groups=rules_lint_report"
+	"--output_groups=rules_lint_output"
 	"--remote_download_regex='.*AspectRulesLint.*'"
 )
 
@@ -58,6 +58,11 @@ if [ $1 == "--fail-on-violation" ]; then
 	)
 	shift
 fi
+
+# Allow a `--fix` option on the command-line.
+# This happens to make output of the linter such as ruff's
+# [*] 1 fixable with the `--fix` option.
+# so that the naive thing of pasting that flag to lint.sh will do what the user expects.
 if [ $1 == "--fix" ]; then
 	fix="patch"
 	args+=(
@@ -78,32 +83,29 @@ bazel build ${args[@]} $@
 # TODO: Maybe this could be hermetic with bazel run @aspect_bazel_lib//tools:jq or sth
 if [ $machine == "Windows" ]; then
     # jq on windows outputs CRLF which breaks this script. https://github.com/jqlang/jq/issues/92
-    valid_reports=$(jq --arg ext report --raw-output "$filter" "$buildevents" | tr -d '\r')
+    valid_outputs=$(jq --arg ext .txt --raw-output "$filter" "$buildevents" | tr -d '\r')
 else
-    valid_reports=$(jq --arg ext report --raw-output "$filter" "$buildevents")
+    valid_outputs=$(jq --arg ext .txt --raw-output "$filter" "$buildevents")
 fi
 
 # Show the results.
-while IFS= read -r report; do
-	# Exclude coverage reports, and check if the report is empty.
-	if [[ "$report" == *coverage.dat ]] || [[ ! -s "$report" ]]; then
+while IFS= read -r output; do
+	# Exclude coverage reports, and check if the output is empty.
+	if [[ "$output" == *coverage.dat ]] || [[ ! -s "$output" ]]; then
 		# Report is empty. No linting errors.
 		continue
 	fi
-	echo "From ${report}:"
-	cat "${report}"
+	echo "From ${output}:"
+	cat "${output}"
 	echo
-done <<<"$valid_reports"
+done <<<"$valid_outputs"
 
-# This happens to make output of the linter such as ruff's
-# [*] 1 fixable with the `--fix` option.
-# so that the naive thing of pasting that flag to lint.sh will do what the user expects.
 if [ -n "$fix" ]; then
-	valid_patches=$valid_reports
+	valid_patches=$valid_outputs
 	while IFS= read -r patch; do
-		# Exclude coverage reports, and check if the report is empty.
+		# Exclude coverage, and check if the patch is empty.
 		if [[ "$patch" == *coverage.dat ]] || [[ ! -s "$patch" ]]; then
-			# Report is empty. No linting errors.
+			# Patch is empty. No linting errors.
 			continue
 		fi
