+++ lint/eslint.bzl
@@ -59,14 +59,14 @@ load("//lint/private:lint_aspect.bzl", "LintOptionsInfo", "dummy_successful_lint
 
 _MNEMONIC = "AspectRulesLintESLint"
 
-def _gather_inputs(ctx, srcs):
+def _gather_inputs(ctx, srcs, files):
     inputs = copy_files_to_bin_actions(ctx, srcs)
 
     # Add the config file along with any deps it has on npm packages
     if "gather_files_from_js_providers" in dir(js_lib_helpers):
         # rules_js 1.x
         js_inputs = js_lib_helpers.gather_files_from_js_providers(
-            ctx.attr._config_files + [ctx.attr._workaround_17660, ctx.attr._formatter],
+            ctx.attr._config_files + files,
             include_transitive_sources = True,
             include_declarations = True,
             include_npm_linked_packages = True,
@@ -74,7 +74,7 @@ def _gather_inputs(ctx, srcs):
     else:
         # rules_js 2.x
         js_inputs = js_lib_helpers.gather_files_from_js_infos(
-            ctx.attr._config_files + [ctx.attr._workaround_17660, ctx.attr._formatter],
+            ctx.attr._config_files + files,
             include_sources = True,
             include_transitive_sources = True,
             include_types = True,
@@ -84,7 +84,7 @@ def _gather_inputs(ctx, srcs):
     inputs.extend(js_inputs.to_list())
     return inputs
 
-def eslint_action(ctx, executable, srcs, report, exit_code = None):
+def eslint_action(ctx, executable, srcs, stdout, exit_code = None, format = None, env = {}):
     """Create a Bazel Action that spawns an eslint process.
 
     Adapter for wrapping Bazel around
@@ -94,46 +94,59 @@ def eslint_action(ctx, executable, srcs, report, exit_code = None):
         ctx: an action context OR aspect context
         executable: struct with an eslint field
         srcs: list of file objects to lint
-        report: output file containing the stdout or --output-file of eslint
+        stdout: output file containing the stdout or --output-file of eslint
         exit_code: output file containing the exit code of eslint.
             If None, then fail the build when eslint exits non-zero.
+        format: value for eslint `--format` CLI flag
+        env: additional environment variables
     """
 
+    format = format or ctx.attr._compact_formatter
+
     args = ctx.actions.args()
+    file_inputs = [ctx.attr._workaround_17660]
 
     # TODO: enable if debug config, similar to rules_ts
     # args.add("--debug")
-
-    args.add_all(["--format", "../../../" + ctx.file._formatter.path])
+    if type(format) == "string":
+        args.add_all(["--format", format])
+    else:
+        args.add_all(["--format", "../../../" + format.files.to_list()[0].path])
+        file_inputs.append(format)
     args.add_all([s.short_path for s in srcs])
 
-    env = {"BAZEL_BINDIR": ctx.bin_dir.path}
+    if exit_code == "discard":
+        exit_code = ctx.actions.declare_file("_discard_eslint_exit_code", sibling = stdout)
 
     if not exit_code:
         ctx.actions.run_shell(
-            inputs = _gather_inputs(ctx, srcs),
-            outputs = [report],
+            inputs = _gather_inputs(ctx, srcs, file_inputs),
+            outputs = [stdout],
             tools = [executable._eslint],
             arguments = [args],
-            command = executable._eslint.path + " $@ && touch " + report.path,
-            env = env,
+            command = executable._eslint.path + " $@ && touch " + stdout.path,
+            env = dict(env, **{
+                "BAZEL_BINDIR": ctx.bin_dir.path,
+            }),
             mnemonic = _MNEMONIC,
             progress_message = "Linting %{label} with ESLint",
         )
     else:
-        # Workaround: create an empty report file in case eslint doesn't write one
+        # Workaround: create an empty file in case eslint doesn't write one
         # Use `../../..` to return to the execroot?
         args.add_joined(["--node_options", "--require", "../../../" + ctx.file._workaround_17660.path], join_with = "=")
 
-        args.add_all(["--output-file", report.short_path])
-        env["JS_BINARY__EXIT_CODE_OUTPUT_FILE"] = exit_code.path
+        args.add_all(["--output-file", stdout.short_path])
 
         ctx.actions.run(
-            inputs = _gather_inputs(ctx, srcs),
-            outputs = [report, exit_code],
+            inputs = _gather_inputs(ctx, srcs, file_inputs),
+            outputs = [stdout, exit_code],
             executable = executable._eslint,
             arguments = [args],
-            env = env,
+            env = dict(env, **{
+                "BAZEL_BINDIR": ctx.bin_dir.path,
+                "JS_BINARY__EXIT_CODE_OUTPUT_FILE": exit_code.path,
+            }),
             mnemonic = _MNEMONIC,
             progress_message = "Linting %{label} with ESLint",
         )
@@ -155,7 +168,7 @@ def eslint_fix(ctx, executable, srcs, patch, stdout, exit_code):
         output = patch_cfg,
         content = json.encode({
             "linter": executable._eslint.path,
-            "args": ["--fix", "--format", "../../../" + ctx.file._formatter.path] + [s.short_path for s in srcs],
+            "args": ["--fix", "--format", "../../../" + ctx.file._compact_formatter.path] + [s.short_path for s in srcs],
             "env": {"BAZEL_BINDIR": ctx.bin_dir.path},
             "files_to_diff": [s.path for s in srcs],
             "output": patch.path,
@@ -163,7 +176,7 @@ def eslint_fix(ctx, executable, srcs, patch, stdout, exit_code):
     )
 
     ctx.actions.run(
-        inputs = _gather_inputs(ctx, srcs) + [patch_cfg],
+        inputs = _gather_inputs(ctx, srcs, [ctx.attr._workaround_17660, ctx.attr._compact_formatter]) + [patch_cfg],
         outputs = [patch, stdout, exit_code],
         executable = executable._patcher,
         arguments = [patch_cfg.path],
@@ -184,19 +197,27 @@ def _eslint_aspect_impl(target, ctx):
         return []
 
     files_to_lint = filter_srcs(ctx.rule)
+    output = None
 
     if ctx.attr._options[LintOptionsInfo].fix:
-        patch, report, exit_code, info = patch_and_report_files(_MNEMONIC, target, ctx)
+        patch, output, report, exit_code, info = patch_and_report_files(_MNEMONIC, target, ctx)
         if len(files_to_lint) == 0:
             dummy_successful_lint_action(ctx, report, exit_code, patch)
         else:
             eslint_fix(ctx, ctx.executable, files_to_lint, patch, report, exit_code)
     else:
-        report, exit_code, info = report_files(_MNEMONIC, target, ctx)
+        output, report, exit_code, info = report_files(_MNEMONIC, target, ctx)
         if len(files_to_lint) == 0:
-            dummy_successful_lint_action(ctx, report, exit_code)
+            dummy_successful_lint_action(ctx, output, exit_code)
         else:
-            eslint_action(ctx, ctx.executable, files_to_lint, report, exit_code)
+            eslint_action(ctx, ctx.executable, files_to_lint, output, exit_code, format = "stylish", env = {
+                # https://www.npmjs.com/package/chalk#chalklevel
+                # Force 256 color support even when a tty isn't detected
+                "FORCE_COLOR": "2",
+            })
+
+    if report:
+        eslint_action(ctx, ctx.executable, files_to_lint, report, exit_code = "discard")
 
     return [info]
 
@@ -243,8 +264,8 @@ def lint_eslint_aspect(binary, configs, rule_kinds = ["js_library", "ts_project"
                 allow_single_file = True,
                 cfg = "exec",
             ),
-            "_formatter": attr.label(
-                default = "@aspect_rules_lint//lint:eslint.bazel-formatter",
+            "_compact_formatter": attr.label(
+                default = "@aspect_rules_lint//lint:eslint.compact-formatter",
                 allow_single_file = True,
                 cfg = "exec",
             ),
